+++ src/tests/Outlets.spec.ts
@@ -1,10 +1,11 @@
-
 import { ShellSdk } from '../ShellSdk';
 import { SHELL_EVENTS } from '../ShellEvents';
 import * as sinon from 'sinon';
 
-describe('Outlets', () => {
+const EXTENSION_SRC = 'http://local.coresystems.net/extension';
+const EXTENSION_ORIGIN = 'http://local.coresystems.net';
 
+describe('Outlets', () => {
   let sdk: ShellSdk;
   let sdkTarget: any;
   let sdkOrigin: string;
@@ -20,13 +21,13 @@ describe('Outlets', () => {
       addEventListener: (eventType, callback) => {
         windowMockEventListenerAdded = true;
         windowMockEventType = eventType;
-        windowMockCallback = callback
-      }
+        windowMockCallback = callback;
+      },
     };
 
     sdkOrigin = 'fsm-sdk.net';
     sdkTarget = {
-      postMessage: sinon.stub()
+      postMessage: sinon.stub(),
     };
   });
 
@@ -37,12 +38,12 @@ describe('Outlets', () => {
 
     // postMessage catch messages send to outlets
     const postMessage = sinon.spy();
-    const iframe = {
-      contentWindow: {
-        postMessage
-      } as any as Window
-    } as any as HTMLIFrameElement;
-    
+    const iframe = ({
+      contentWindow: ({
+        postMessage,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement;
+
     sdk.registerOutlet(iframe);
     sdk.unregisterOutlet(iframe);
 
@@ -51,10 +52,10 @@ describe('Outlets', () => {
         type: SHELL_EVENTS.Version1.SET_VIEW_STATE,
         value: {
           key: 'TECHNICIAN',
-          value: 42
+          value: 42,
         },
-        to: ['abc'] // `to` parameter should be ignored
-      }
+        to: ['abc'], // `to` parameter should be ignored
+      },
     });
 
     expect(postMessage.called).toBe(false);
@@ -74,29 +75,29 @@ describe('Outlets', () => {
       origin = _origin;
     });
 
-    sdk.onViewState('TECHNICIAN', id => technicianId = id);
+    sdk.onViewState('TECHNICIAN', (id) => (technicianId = id));
 
     // Define two outlets with different frame
-    sdk.registerOutlet({
-      contentWindow: {
-        postMessage
-      } as any as Window
-    } as any as HTMLIFrameElement);
-    sdk.registerOutlet({
-      contentWindow: {
-        postMessage
-      } as any as Window
-    } as any as HTMLIFrameElement);
+    sdk.registerOutlet(({
+      contentWindow: ({
+        postMessage,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement);
+    sdk.registerOutlet(({
+      contentWindow: ({
+        postMessage,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement);
 
     windowMockCallback({
       data: {
         type: SHELL_EVENTS.Version1.SET_VIEW_STATE,
         value: {
           key: 'TECHNICIAN',
-          value: 42
+          value: 42,
         },
-        to: ['abc'] // `to` parameter should be ignored
-      }
+        to: ['abc'], // `to` parameter should be ignored
+      },
     });
 
     expect(type).toEqual(SHELL_EVENTS.Version1.SET_VIEW_STATE);
@@ -120,9 +121,9 @@ describe('Outlets', () => {
         type: SHELL_EVENTS.Version1.SET_VIEW_STATE,
         value: {
           key: 'TECHNICIAN',
-          value: 42
-        }
-      }
+          value: 42,
+        },
+      },
     });
 
     expect(event.called).toBe(false);
@@ -135,9 +136,9 @@ describe('Outlets', () => {
         type: SHELL_EVENTS.Version1.SET_VIEW_STATE,
         value: {
           key: 'TECHNICIAN',
-          value: 42
-        }
-      }
+          value: 42,
+        },
+      },
     });
     expect(event.called).toBe(false);
   });
@@ -152,17 +153,17 @@ describe('Outlets', () => {
 
     // postMessage catch messages send to outlets
     const postMessage = sinon.spy();
-    sdk.registerOutlet({
-      contentWindow: {
-        postMessage
-      } as any as Window
-    } as any as HTMLIFrameElement);
+    sdk.registerOutlet(({
+      contentWindow: ({
+        postMessage,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement);
 
     windowMockCallback({
       data: {
         type: SHELL_EVENTS.Version1.TO_APP,
-        value: 'HI APP'
-      }
+        value: 'HI APP',
+      },
     });
 
     expect(value).toEqual('HI APP');
@@ -170,30 +171,35 @@ describe('Outlets', () => {
   });
 
   it('should not handle TO_APP message from an outlet but send parent', () => {
-    
     const postMessageParent = sinon.spy();
-    sdk = ShellSdk.init({
-      postMessage: postMessageParent
-    } as any as Window, sdkOrigin, windowMock);
+    sdk = ShellSdk.init(
+      ({
+        postMessage: postMessageParent,
+      } as any) as Window,
+      sdkOrigin,
+      windowMock
+    );
 
     let handleMessage = sinon.spy();
     sdk.on(SHELL_EVENTS.Version1.TO_APP, handleMessage);
 
     // postMessage catch messages send to outlets
     const postMessageOutlet = sinon.spy();
-    const iframe = {
-      contentWindow: {
-        postMessage: postMessageOutlet
-      } as any as Window
-    } as any as HTMLIFrameElement;
+    const iframe = ({
+      src: EXTENSION_SRC,
+      contentWindow: ({
+        postMessage: postMessageOutlet,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement;
     sdk.registerOutlet(iframe);
 
     windowMockCallback({
       source: iframe.contentWindow,
+      origin: EXTENSION_ORIGIN,
       data: {
         type: SHELL_EVENTS.Version1.TO_APP,
-        value: 'RANDOM_VALUE'
-      }
+        value: 'RANDOM_VALUE',
+      },
     });
 
     expect(postMessageParent.called).toBe(true);
@@ -202,69 +208,81 @@ describe('Outlets', () => {
   });
 
   it('should ignore SET_VIEW_STATE message from an outlet for security reason', () => {
-
     const postMessageParent = sinon.spy();
-    sdk = ShellSdk.init({
-      postMessage: postMessageParent
-    } as any as Window, sdkOrigin, windowMock);
+    sdk = ShellSdk.init(
+      ({
+        postMessage: postMessageParent,
+      } as any) as Window,
+      sdkOrigin,
+      windowMock
+    );
 
     let handleMessage = sinon.spy();
     sdk.on(SHELL_EVENTS.Version1.SET_VIEW_STATE, handleMessage);
 
     // postMessage catch messages send to outlets
     const postMessageOutlet = sinon.spy();
-    const iframe = {
-      contentWindow: {
-        postMessage: postMessageOutlet
-      } as any as Window
-    } as any as HTMLIFrameElement;
+    const iframe = ({
+      src: EXTENSION_SRC,
+      contentWindow: ({
+        postMessage: postMessageOutlet,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement;
     sdk.registerOutlet(iframe);
 
     const consoleSpy = sinon.spy(console, 'warn');
 
     windowMockCallback({
       source: iframe.contentWindow,
+      origin: EXTENSION_ORIGIN,
       data: {
         type: SHELL_EVENTS.Version1.SET_VIEW_STATE,
         value: {
           key: 'TECHNICIAN',
-          value: 42
-        }
-      }
+          value: 42,
+        },
+      },
     });
 
-    expect(consoleSpy.calledWith('[ShellSDk] A plugin tried to update viewState using SetViewState which is not allowed for security reason.')).toBe(true);
+    expect(
+      consoleSpy.calledWith(
+        '[ShellSDk] A plugin tried to update viewState using SetViewState which is not allowed for security reason.'
+      )
+    ).toBe(true);
     expect(postMessageParent.called).toBe(false);
     expect(handleMessage.called).toBe(false);
     expect(postMessageOutlet.called).toBe(false);
   });
 
   it('should outlet send to parent loading_success on require_context', () => {
-    
     const postMessageParent = sinon.spy();
-    sdk = ShellSdk.init({
-      postMessage: postMessageParent
-    } as any as Window, sdkOrigin, windowMock);
+    sdk = ShellSdk.init(
+      ({
+        postMessage: postMessageParent,
+      } as any) as Window,
+      sdkOrigin,
+      windowMock
+    );
 
     let handleMessage = sinon.spy();
     sdk.on(SHELL_EVENTS.Version1.REQUIRE_CONTEXT, handleMessage);
 
     // postMessage catch messages send to outlets
     const postMessageOutlet = sinon.spy();
-    const iframe = {
-      contentWindow: {
-        postMessage: postMessageOutlet
-      } as any as Window
-    } as any as HTMLIFrameElement;
+    const iframe = ({
+      contentWindow: ({
+        postMessage: postMessageOutlet,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement;
     sdk.registerOutlet(iframe);
 
     windowMockCallback({
       data: {
         type: SHELL_EVENTS.Version1.REQUIRE_CONTEXT,
         value: {
-          message: 'test'
-        }
-      }
+          message: 'test',
+        },
+      },
     });
 
     expect(postMessageParent.called).toBe(true);
@@ -279,32 +297,32 @@ describe('Outlets', () => {
 
     // postMessage catch messages send to outlets
     const postMessage = sinon.spy();
-    sdk.registerOutlet({
-      contentWindow: {
-        postMessage
-      } as any as Window
-    } as any as HTMLIFrameElement);
+    sdk.registerOutlet(({
+      contentWindow: ({
+        postMessage,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement);
 
     windowMockCallback({
       data: {
         type: SHELL_EVENTS.Version1.OUTLET.REQUEST_CONTEXT,
         value: {
           isConfigurationMode: true,
-          target: ''
-        }
-      }
+          target: '',
+        },
+      },
     });
 
     expect(postMessage.called).toBe(false);
     postMessage.resetHistory();
-    
+
     windowMockCallback({
       data: {
         type: SHELL_EVENTS.Version1.OUTLET.REQUEST_CONTEXT,
         value: {
-          isConfigurationMode: true
-        }
-      }
+          isConfigurationMode: true,
+        },
+      },
     });
 
     expect(postMessage.called).toBe(true);
@@ -315,43 +333,48 @@ describe('Outlets', () => {
         type: SHELL_EVENTS.Version1.OUTLET.REQUEST_CONTEXT,
         value: {
           isConfigurationMode: true,
-          plugin: {}
-        }
-      }
+          plugin: {},
+        },
+      },
     });
 
     expect(postMessage.called).toBe(false);
-
   });
 
-
   it('should return SHELL_EVENTS.Version1.OUTLET.LOADING_FAIL if reached maximum depth', () => {
-
     const postMessageParent = sinon.spy();
-    sdk = ShellSdk.init({
-      postMessage: postMessageParent
-    } as any as Window, sdkOrigin, windowMock, null, 3);
+    sdk = ShellSdk.init(
+      ({
+        postMessage: postMessageParent,
+      } as any) as Window,
+      sdkOrigin,
+      windowMock,
+      null,
+      3
+    );
 
     let handleMessage = sinon.spy();
     sdk.on(SHELL_EVENTS.Version1.OUTLET.LOADING_FAIL, handleMessage);
 
     const postMessageOutlet = sinon.spy();
-    const iframe = {
-      contentWindow: {
-        postMessage: postMessageOutlet
-      } as any as Window
-    } as any as HTMLIFrameElement;
+    const iframe = ({
+      src: EXTENSION_SRC,
+      contentWindow: ({
+        postMessage: postMessageOutlet,
+      } as any) as Window,
+    } as any) as HTMLIFrameElement;
     sdk.registerOutlet(iframe);
 
     windowMockCallback({
       source: iframe.contentWindow,
+      origin: EXTENSION_ORIGIN,
       data: {
         type: SHELL_EVENTS.Version1.OUTLET.REQUEST_CONTEXT,
         value: {
-          target: 'test'
+          target: 'test',
         },
-        from: ['a', 'b']
-      }
+        from: ['a', 'b'],
+      },
     });
 
     expect(postMessageParent.called).toBe(true);
@@ -361,17 +384,17 @@ describe('Outlets', () => {
 
     windowMockCallback({
       source: iframe.contentWindow,
+      origin: EXTENSION_ORIGIN,
       data: {
         type: SHELL_EVENTS.Version1.OUTLET.REQUEST_CONTEXT,
         value: {
-          target: 'test'
+          target: 'test',
         },
-        from: ['a', 'b', 'c']
-      }
+        from: ['a', 'b', 'c'],
+      },
     });
 
     expect(postMessageParent.called).toBe(false);
     expect(postMessageOutlet.called).toBe(true);
   });
-
 });