+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2021-01-19
+
+### Update
+
+- Validate iFrame origing against postMessage event origin when proxing events from outlets
+- Added ability to dynamically add/remove allowed origins
+
 ## [1.7.3] - 2020-12-11
 
 ### Update